Feat/screentime: 존재하지 않는 스크린 타임 조회에 대한 에러 응답 폐기

+++ src/main/java/dev/wetox/WetoxRESTful/screentime/CategoryScreenTime.java
@@ -23,7 +23,7 @@ public class CategoryScreenTime {
     @Enumerated(STRING)
     private Category category;
 
-    private Double duration;
+    private Long duration;
 
     public static CategoryScreenTime build(ScreenTime screenTime, CategoryScreenTimeRequest categoryScreenTimeRequest) {
         CategoryScreenTime categoryScreenTime = new CategoryScreenTime();
